# configured MqttRepository.kt to coroutines and flow.

+++ app/src/main/java/com/vovan/diplomaapp/data/dto/DtoClasses.kt
@@ -10,6 +10,6 @@ data class SensorsDTO(
     val time: Long,
     val device: String,
     val temperature: Float,
-    val light: Int,
+    val light: Float,
     val pressure: Int
 )
